Update EnemyMechanics.cs
Added a new variable to allow to distinguish between regular wait time and damaged wait time.
Whenever an enemy switches states, a new roam position will be allocated to allow the game to be more dynamic and so that the enemy won't go halfway across the map after chasing the player.
If the enemy sees the player, it will no longer keep waiting.